requirements.txt
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.1
numpy==1.22.3
pandas==1.4.2
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
